requirements.txt
crewai
crewai-tools
openai
langchain-openai
python-dotenv
pydantic
